Big updates to the TestGame.
TestGame now has player controls, combat, basic enemy AI, and more.  This is essentially the completed version of the game from the Night Run Studios youtube tutorial series.  We will see how much of it will be useful, but the EnemyCombat script and EnemyMovement script will likely be carried over in some fashion.